Add files via upload
This repository showcases an ETL pipeline created with Apache Spark which manages JSON files. 

Spark is an open-source distributed general-purpose cluster-computing framework, that provides an interface for programming entire clusters with implicit data parallelism and fault tolerance. A unified analytics engine for large-scale data processing, a framework which is used for processing, querying and analyzing Big data. 

Computation is done in memory hence it’s multiple-fold fasters than the competitors like MapReduce and others with the rate at which terabytes of data is being produced every day, there was a need for a solution that could provide real-time analysis at high speed. Some of the Spark features are: